Add NoOpCapturedOutput class for non-capturing output scenarios
