+++ vedro/core/output_capturer/__init__.py
@@ -1,6 +1,7 @@
-from ._captured_output import CapturedOutput
+from ._captured_output import CapturedOutput, NoOpCapturedOutput
 from ._output_capturer import OutputCapturer
 from ._stream_buffer import StreamBuffer
 from ._stream_view import StreamView
 
-__all__ = ("CapturedOutput", "StreamBuffer", "StreamView", "OutputCapturer",)
+__all__ = ("CapturedOutput", "StreamBuffer", "StreamView", "OutputCapturer",
+           "NoOpCapturedOutput",)
